# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiTokenCreateInput": {
        "additionalProperties": false,
        "properties": {
          "expires_days": {
            "anyOf": [
              {
                "maximum": 3650.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires Days"
          },
          "name": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "viewer",
              "admin"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "name",
          "role"
        ],
        "title": "ApiTokenCreateInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "POps central server API. `/api/v1` is the stable surface for integrations; the same endpoints answer under plain `/api` for the panel and existing agents. Authenticate with an API token (`Authorization: Bearer pops_...`, created by a superadmin in Ayarlar → Güvenlik) or a panel login JWT. Paths marked deprecated still work; use their REST names. Reference: docs/api.md.",
    "title": "POps API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tokens": {
      "get": {
        "description": "Bütün jetonlar, yeniden eskiye (iptal edilmiş ve süresi dolmuşlar da). Jetonun kendisi hiçbir zaman dönmez.",
        "operationId": "list_tokens_api_tokens_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Tokens",
        "tags": [
          "tokens"
        ]
      },
      "post": {
        "operationId": "create_token_api_tokens_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiTokenCreateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Token",
        "tags": [
          "tokens"
        ]
      }
    },
    "/api/v1/tokens/{token_id}": {
      "delete": {
        "description": "İptal anında geçerlidir: jeton her istekte veritabanına karşı denetlenir. Kayıt silinmez (kim, ne zaman).",
        "operationId": "revoke_token_api_tokens__token_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "title": "Token Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Token",
        "tags": [
          "tokens"
        ]
      }
    },
...
  }
}